requirements.txt
absl-py==0.11.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.3.4
asn1crypto==0.24.0
astor==0.8.1
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
blis==0.7.4
Brlapi==0.6.6
cached-property==1.5.2
catalogue==1.0.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==8.0.1
colorama==0.4.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.5
Cython==0.29.23
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
Django==3.2.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
ffmpeg==1.4
filelock==3.0.12
future==0.18.2
gast==0.4.0
grpcio==1.35.0
h5py==3.1.0
httplib2==0.9.2
idna==2.6
image==1.5.33
imageio==2.9.0
imageio-ffmpeg==0.4.4
importlib-metadata==3.7.2
importlib-resources==4.1.1
ipykernel==5.5.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.7.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
line-profiler==2.1
llvmlite==0.34.0
louis==3.5.0
lxml==4.5.2
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
mock==4.0.3
module==0.0.4
moviepy==1.0.3
murmurhash==1.0.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.5.1
nltk==3.6.2
notebook==6.2.0
numpy==1.19.5
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
opencv-contrib-python==4.0.0.21
opencv-python==4.1.2.30
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.2.1
pickleshare==0.7.5
Pillow==8.1.2
pipenv==2020.11.15
plac==1.1.3
preshed==3.0.5
proglog==0.1.9
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.14.0
ptyprocess==0.7.0
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pygame==2.0.1
pygifsicle==1.0.4
Pygments==2.8.1
pygobject==3.26.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pyrsistent==0.17.3
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2018.3
PyWavelets==1.1.1
pywinpty==0.5.7
pyxdg==0.25
PyYAML==3.12
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
regex==2021.4.4
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
ruamel.yaml==0.11.14
scikit-image==0.17.2
scipy==1.5.4
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.13.2
six==1.15.0
spacy==2.3.5
sqlparse==0.4.1
srsly==1.0.5
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
tf-bodypix==0.3.5
thinc==7.4.5
tifffile==2020.9.3
toml==0.10.2
tornado==6.1
tqdm==4.55.0
traitlets==4.3.3
typed-ast==1.4.2
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==20.2.2
virtualenv-clone==0.5.4
wadllib==1.3.2
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
zipp==3.4.1
zope.interface==4.3.2
